Link header username label to edit dialog (#3762)
Cursor pointer added to user tab in header.
Page title modified in base and user _list twig template.
Table click function moved to main coffee from list coffee
so that the function can be accessed in the header.
Show alert modified in list coffee so that it extend the
notes data property and can be accessed from the main coffee.