Employee should be able to add only own travel request (#3796)
Created validation so that when the user tries to
modify the employee id while creating or editing a
travel request an error appears on top of the page.
Created function setTravelRequestForm because
the form needs to be renewed after false user was
entered.
Modified travel type so that it disables the employee
name field if user is not admin.