Create travel expense from travel request list (#3888)
1. From expense controller travel request id, arrival and
departure dates are passed to the twig template.
2. Redirecting changed in expense controller so it wont
redirect to travel expense listing after it has been
created.
3. showTravelExpense template changed so now it display the
id of the travel request and after form is submitted
it passes the travel request id so it can be added
to the database.
4. Travel request listing page modified so now it contains an
options to add or edit the travel expense of the created
travel requests.
5. Read only attribute added to arrival and departure date
in expense form, and red-text class has been added to change
the color of the travel request id in the travel expense
form.