Create paging for listing pages (#4187)
1) Added methods init pager and travel request
listing to re initialize the listeners in them when
a new page gets loaded.
2) Created travel and user bundle pager max results and max
pages to show in config files in parameters.yml.dist in
the main app folder.
3) In travel controller search action was moved into
listing action.
4) In travel request repository search action was modified
so its shows results according to the rights the
user has got to the request.
5) Created get pagination function for paging on
the travel request listing page.
6) Force tr status method created to be able to add a
status to a travel request without sending a notificaiton
or an email.
7) Created new method in travel request service to
set access right to the travel requests on the listing page.
8) Created search button listener in travel request
list coffee file.
9) Extended travel request and user partial listing
pages to include pager.
10) Method form field is empty in main coffee file was
changed to ignore hidden input fields.
11) Pager partial template file was created.