FIX: Print view missing rates bugfix
Rates could not be fetched for expenses having
the for approval status as times where set and
not matching in mysql query's.

MySql versions should be checked on this type of
feature to see if times '00:00:00' will always be
recognized as not set and only matched on the date
part.