Notification container usability enhancements (#4376)
The display behavior is changed to display the bell
icon only partly to decrease its current state's importance.
When hovering it slides out to its normal position.
Once new alerts are received the icon color changes, the count
is updated as before and the position state changed to be
fully visible.