FIX: Travel request status change reload bug
Whenever the change status action failed due to an
uncatched exception or fatal error, the list page
reloaded obscuring the current happening.
The ajax action is changed to handle this case blocking
the "complete" handler from reloading the page.

TODO: A decent exception message has to be shown to the user
and the status select reset.