FIX: Travel list page acl and button text for GM (#4456, #4458, #4459)
If the admin wanted to see the travel list page then it given an error.
Because the teIds template variable in the twig had not stausName
element. So in this case the statusName was not available.
In the Travel request service, the setTravelRequestListingRights was
extended to add statusName element to the teIds array.