FIX: Travel request list change status logic (#4454)
The statuses' drop down list generator loop was modified in the Travel
request's partial list template.
The getNextAvailableStates was modifed in the Travel request service
class. If the current user is not the general manager then this method
will exclude the approved, rejected, revise states from the next states
list.