Time sheet PDF generation (#4892)
The path of the CSS files in the print view of travel expense and
time sheet is modified in order to cut off the end of assets if it
has got queries (which is the asset version). Because the TCPDF
cannot handle the assets with the query in this case it will not be able
to use the external css files.

1) The exportTimeSheetToPDFAction is created to handle generate time sheet
   in pdf.
2) The PDF manager's constructor is modified to be able to set the
   type and page size of the pdf. Type cans be landscape or portrait.
   Page size cans be A4, A5, etc.
3) The print view of travel expense is modified.
   The download pdf icon is replaced with a fontawesome icon.
   The link of the css file is modified in order to the TCPDF cans read in
   the external css file.