FIX: Asset cache control feature (#4923)
The asset methods is removed in front of asset_url.
In Production environment only the asset_url variable generate out
the asset version too.
But in the Development environment it does not generate the asset versions
(It does only with the asset methods in dev environment).
However in Production environment the asset method with asset_url parameter
will generate the asset version twice in the query!