System leave categories (#5203)
Run Migration! And you can load the fixtures.
Fixed ldap enabled in User: setting the default false value for it
in the constructor of User entity in UserBundle.

System property was added to the AbstractBase to handle the system flag
properties for that entities which extend this abstract class.
The list ordering of leave categories in admin leave controller changed
to generate the system categories on the top of list page. And preventing
the deletion of system categories.
Fixtures was exteneded with this system flag.