FIX: Create LR Group entity for the mass LRs (#5251)
Migration file!
Created a flag for massive leave requests to identify them from the
employee leave requests in the leave request entity.
In the controller a new massive leave requests is created.
The leave request service is modified to set the flag for massiave lrs.
In the list page the massive leave requests have got a group icon, italic
and weighted name, and the status for them removed and not editable.