Pager shows zeroth page in the list. (#5349)
setPagerNumbering method is modified in the list coffee.
The minimum value of interval will set if the actual offset is not reach
the max pages value.
If the actual offset is reach or higher than the max pages values then
in this case the min value will increase.
There is a special case when the max pages value is an odd number. Because
in this case the current page number cannot be placed at the middle, and
the difference of max and min value always have to be audited in order to
not be greater than the max pages value.