Leave request group ordering (#5610)
The query builder has limitations so in this case the repository was not
extended with a new query. Because in the select there is a different
column (not part of the entity) as an alias then in the result it will be
an other element. The first is the entity and the second is the alias.

1) initPager function is refactorized. It uses only one ajax calling with
   the ordering parameters.
2) The list action is modified in the leave controller to set the parent
   leave id for the leave requests (their mass leave request id).
3) In the Leave request entity the parentLeaveRequestId runtime
   property is added to store its parent leave request id.
4) findAllByFiltersPaginated method in the leave request repository
   is modified. The order parameters changed.
5) parentLeaveRequestId runtime property is used now.