Create leave request for state not remembered on errors (#5613)
The show leave request action is modified n the leave controller.
The selected employees is passed to the view and the requestFor variable
is initialized with the suitable value instead of null.
In the show leave request template suitable display classes added on the
employee list and employee input.
In the show leave request coffee file the request for form part
is refactored.