Time sheet adds next month first day (#5675)
The end date is modified in the getTimeSheetPage method of timesheet
controller to get the last day of this month. (Because it added 1 month)
In the time sheet print view the br is removed, because in the print view
generation it pushed down the first table.
Time sheet print view scss is modified. The h3 get higher top margin value.
CSS file is complied for the TCPDF plugin.