Change the currency rate controller to rest controller (#5765)
The currency rate controller is changed to rest controller. The actions
intensively use the query param annotation. You can read after in the
wiki.
Test files are changed.
In the findAllByDates() method of rate repository is modified. The start and
end dates' times are initialized as 0. It was necessary to find the today's
rate, because in some case the start date did not start with 0 hour, minute.