FIX: Give Product name and implement throughout application (5545)
In the status manager the bundle name splitting failed. Because the
splitting rule is based on the capital letters. Before the changing
the Notes contained only one capital letter and now the OpitHrm contains
two capital letters so after the splitting the number of pieces is increased
with one. So the index of pieces is increased to fetch the bundle name
in the right way.