Secured area firewall access control changes (#5942)
This commit is for preparation of secure channels
on access control for firewalls.
The security for login pages is added back to the
firewall and requires now anonymous access.

To add secured channels for implementations, local commits
have to be created on release branches adding
required_channel: https to the access control lists.