Job title list does not refresh after deleting (#5978)
In the app's main coffee the buttons are modified of delete actions.
The yes button call the destroy method. The destroy method will remove
the dialog. The no button call the close event and then the close method
will run and after it will destroy the dialog too.
The list coffee importing is removed from the job title list page. Because
It is specified for the user list page.