Team leave calendar enhancements (#5981)
1) Created two new css classes.
2) Created new function in leave date repository two
find all leave dates in a date range.
3) Refactored calendar controller showTeamEmployeeLeavesAction
to add the leave days in a date range to the
array returned.
4) Team leaves calendar coffee file modified to start
calendar with Monday.
Event after rendering function added to check if an
event has a specific class and if it does set the
background color of its column in the calendar.