Timesheet PDF missing border bugfix (#6154)
Each table does now have a border at the bottom, even
if an employee left during the month or it ends with
a weekend.

The tr:last-child class that adds the bottom border is not
recognized by the TCPDF library, a new class is added in
case of pdf exports.